Add sub directories to split up the code
In the near future we want to switch from JS to TS to
develop types for passing messages between the different
components of Zen. The sub-directories will help us create
these communication boundaries.
