+++ lib/aws/aws-bucket-only.template
