+++ lib/aws/aws.template
