+++ lib/aws/lambda.js
