+++ lib/cli/cli.js
@@ -3,11 +3,11 @@
 // Normalize whether the cli is run directly or via node
 if (!process.argv[0].match(/zen$/)) process.argv.shift()
 
-require('./index')
+require('../client/index')
 
 const cmd = process.argv[1] || 'run'
 if (cmd == 'server') {
-  const Server = require('./server')
+  const Server = require('../local-server')
   new Server()
 }
