+++ lib/client/head.js
