+++ lib/client/index.js
@@ -1,8 +1,8 @@
 const path = require('path')
 const https = require('https')
-const Util = require('./util')
+const Util = require('../util')
 const AWS = require('aws-sdk')
-const S3Sync = require('./s3-sync')
+const S3Sync = require('../local-server/s3-sync')
 const Journal = require('./journal')
 const uuidv4 = require('uuid/v4')
 
@@ -10,7 +10,7 @@ require('sugar').extend()
 global.Zen = {}
 
 // load the config with some defaults
-let config = Zen.config = require(path.join(process.cwd(), process.argv[2]))
+let config = Zen.config = require(path.join(process.cwd(), process.argv[2] || "zen.config.js"))
 config.appRoot = path.resolve(process.cwd(), config.appRoot || '')
 config.port = config.port || 3100
 config.testDependencies = config.testDependencies || []
@@ -33,17 +33,17 @@ Zen.journal = new Journal()
 https.globalAgent.maxSockets = 2000 // TODO multiplex over fewer connections
 
 if (config.webpack) { // boot up webpack (if configured)
-  let WebpackAdapter = require('./webpack')
+  let WebpackAdapter = require('../local-server/webpack')
   Zen.webpack = new WebpackAdapter()
 }
 
 Zen.indexHtml = function indexHtml (pageType, forS3) {
-  let deps = ['lib/latte.js']
+  let deps = ['lib/client/latte.js']
   if (pageType == 'head') {
     deps.unshift('icons')
     deps.push('node_modules/svelte/store.umd.js', 'node_modules/fuzzysort/fuzzysort.js', 'svelte/mini.js', 'svelte/command.js')
   }
-  deps.push(`lib/${pageType}.js`) // after Zen dependencies, but before user code
+  deps.push(`lib/client/${pageType}.js`) // after Zen dependencies, but before user code
   let entries = (Zen.webpack && Zen.webpack.compile && Zen.webpack.compile.entrypoints) || []
 
   if (forS3) {
