+++ lib/client/journal.js
@@ -1,5 +1,5 @@
 const path = require('path')
-const util = require('./util')
+const util = require('../util')
 
 module.exports = class Journal {
   constructor () {
