+++ lib/client/webpack-client.js
