+++ lib/client/worker.js
