+++ lib/local-server/index.js
@@ -2,7 +2,7 @@ const path = require('path')
 const http = require('http')
 const connect = require('connect')
 const WebSocket = require('ws')
-const Util = require('./util')
+const Util = require('../util')
 const ChromeWrapper = require('./chrome')
 
 module.exports = class Server {
@@ -19,7 +19,7 @@ module.exports = class Server {
     // start up the local webserver for `head` to connect to
     let app = connect()
     let server = http.createServer(app).listen(Zen.config.port)
-    app.use('/lib', Util.serveWith404(__dirname)) // serve up stuff out of lib
+    app.use('/lib', Util.serveWith404(path.join(__dirname, '..'))) // serve up stuff out of lib
     app.use('/node_modules', Util.serveWith404(path.resolve(Zen.config.appRoot, './node_modules'))) // TODO: this doesn't work when yarn link'd
     app.use('/base', Util.serveWith404(Zen.config.appRoot)) // base serves things out of the application's root
     app.use('/svelte', Util.serveSvelte)
