+++ lib/local-server/s3-sync.js
@@ -4,7 +4,7 @@ const klaw = require('klaw')
 const mime = require('mime-types')
 const AWS = require('aws-sdk')
 const crypto = require('crypto')
-const util = require('./util')
+const util = require('../util')
 
 module.exports = class S3Sync extends EventEmitter {
   constructor() {
@@ -34,7 +34,7 @@ module.exports = class S3Sync extends EventEmitter {
 
     // first, get all the directories we'd like to sync
     let syncedDirs = [
-      {filePath: __dirname, webPath: '/lib', filter: (p) => p.match(/worker.js|latte.js/)}
+      {filePath: path.join(__dirname, '../client'), webPath: '/lib/client', filter: (p) => p.match(/worker.js|latte.js/)}
     ].concat(Zen.config.alsoServe || [])
 
     // recursively stat every directory, getting a list of files
