+++ lib/local-server/webpack.js
@@ -6,7 +6,7 @@ module.exports = class WebpackAdapter extends EventEmitter {
   constructor () {
     super()
     let wcfg = Zen.config.webpack
-    wcfg.entry.bundle.push(path.join(__dirname, 'webpack-client.js'))
+    wcfg.entry.bundle.push(path.join(__dirname, '../client', 'webpack-client.js'))
     wcfg.plugins.push(new webpack.HotModuleReplacementPlugin())
     wcfg.plugins.push(new webpack.ProgressPlugin((pct, message, addInfo) => {
       if (pct > 0 && pct < 1)
