+++ lib/util/index.js
@@ -14,7 +14,7 @@ Util.serveWith404 = function(dir) {
 
 Util.serveSvelte = function(req, res) {
   let name = path.basename(req.url, '.js')
-  fs.readFile(path.join(__dirname, name + '.html'), 'utf8', function (err, data) {
+  fs.readFile(path.join(__dirname, '../client/', name + '.html'), 'utf8', function (err, data) {
     if (err) throw err
     name = name[0].toUpperCase() + name.slice(1)
     try {
@@ -34,7 +34,7 @@ let iconCache = null
 Util.serveIcons = async function(req, res) {
   if (iconCache) return res.end(iconCache)
   let icons = {}
-  let root = path.join(__dirname, '../assets')
+  let root = path.join(__dirname, '../client/assets')
   await Promise.all(fs.readdirSync(root).map(async fname => {
     if (!fname.match(/([\w_\-]+)\.svg$/)) return null
     icons[RegExp.$1.camelize()] = await Util.readFileAsync(path.join(root, fname))
